click outside to hide + install framer motion
